+++ Compiler/Tests/Arithmetic.cs
@@ -6,7 +6,7 @@ namespace Exodrifter.Rumor.Compiler.Tests
 {
 	public static class Arithmetic
 	{
-		#region Literal
+		#region Primitives
 
 		[Test]
 		public static void IntegerSuccess()
@@ -97,6 +97,115 @@ namespace Exodrifter.Rumor.Compiler.Tests
 			);
 		}
 
+		[Test]
+		public static void Binding0Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number);
+
+			var state = new ParserState("foobar()", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind("foobar", () => { return 1234d; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1234),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void Binding1Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number, ValueType.Number);
+
+			var state = new ParserState("foobar(1234)", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double>("foobar", (a) => { return a; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1234),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void Binding2Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(411, 823)", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double>(
+				"foobar", (a, b) => { return a + b; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1234),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void Binding3Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(176, 352, 706)", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double>(
+				"foobar", (a, b, c) => { return a + b + c; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1234),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void Binding4Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(82, 164, 329, 659)", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double, double>(
+				"foobar", (a, b, c, d) => { return a + b + c + d; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1234),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
 		#endregion
 
 		#region Addition
@@ -124,6 +233,115 @@ namespace Exodrifter.Rumor.Compiler.Tests
 			Assert.AreEqual(
 				new NumberValue(1236),
 				n.Evaluate(scope, new RumorBindings())
+			);
+		}
+
+		[Test]
+		public static void AdditionBinding0Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number);
+
+			var state = new ParserState("foobar() + 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind("foobar", () => { return 1234d; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1236),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void AdditionBinding1Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number, ValueType.Number);
+
+			var state = new ParserState("foobar(1234) + 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double>("foobar", (a) => { return a; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1236),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void AdditionBinding2Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(411, 823) + 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double>(
+				"foobar", (a, b) => { return a + b; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1236),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void AdditionBinding3Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(176, 352, 706) + 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double>(
+				"foobar", (a, b, c) => { return a + b + c; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1236),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void AdditionBinding4Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(82, 164, 329, 659) + 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double, double>(
+				"foobar", (a, b, c, d) => { return a + b + c + d; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1236),
+				n.Evaluate(new RumorScope(), bindings)
 			);
 		}
 
@@ -230,6 +448,115 @@ namespace Exodrifter.Rumor.Compiler.Tests
 		}
 
 		[Test]
+		public static void SubtractionBinding0Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number);
+
+			var state = new ParserState("foobar() - 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind("foobar", () => { return 1234d; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1232),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void SubtractionBinding1Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number, ValueType.Number);
+
+			var state = new ParserState("foobar(1234) - 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double>("foobar", (a) => { return a; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1232),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void SubtractionBinding2Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(411, 823) - 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double>(
+				"foobar", (a, b) => { return a + b; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1232),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void SubtractionBinding3Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(176, 352, 706) - 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double>(
+				"foobar", (a, b, c) => { return a + b + c; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1232),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void SubtractionBinding4Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(82, 164, 329, 659) - 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double, double>(
+				"foobar", (a, b, c, d) => { return a + b + c + d; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(1232),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
 		public static void SubtractionPlusSignSuccess()
 		{
 			var state = new ParserState("1-+2", 4);
@@ -332,6 +659,115 @@ namespace Exodrifter.Rumor.Compiler.Tests
 		}
 
 		[Test]
+		public static void MultiplicationBinding0Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number);
+
+			var state = new ParserState("foobar() * 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind("foobar", () => { return 1234d; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(2468),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void MultiplicationBinding1Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number, ValueType.Number);
+
+			var state = new ParserState("foobar(1234) * 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double>("foobar", (a) => { return a; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(2468),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void MultiplicationBinding2Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(411, 823) * 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double>(
+				"foobar", (a, b) => { return a + b; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(2468),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void MultiplicationBinding3Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(176, 352, 706) * 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double>(
+				"foobar", (a, b, c) => { return a + b + c; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(2468),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void MultiplicationBinding4Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(82, 164, 329, 659) * 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double, double>(
+				"foobar", (a, b, c, d) => { return a + b + c + d; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(2468),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
 		public static void MultiplicationPlusSignSuccess()
 		{
 			var state = new ParserState("1*+2", 4);
@@ -430,6 +866,115 @@ namespace Exodrifter.Rumor.Compiler.Tests
 			Assert.AreEqual(
 				new NumberValue(617),
 				n.Evaluate(scope, new RumorBindings())
+			);
+		}
+
+		[Test]
+		public static void DivisionBinding0Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number);
+
+			var state = new ParserState("foobar() / 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind("foobar", () => { return 1234d; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(617),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void DivisionBinding1Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction("foobar", ValueType.Number, ValueType.Number);
+
+			var state = new ParserState("foobar(1234) / 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double>("foobar", (a) => { return a; });
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(617),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void DivisionBinding2Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(411, 823) / 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double>(
+				"foobar", (a, b) => { return a + b; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(617),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void DivisionBinding3Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(176, 352, 706) / 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double>(
+				"foobar", (a, b, c) => { return a + b + c; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(617),
+				n.Evaluate(new RumorScope(), bindings)
+			);
+		}
+
+		[Test]
+		public static void DivisionBinding4Success()
+		{
+			var rps = new RumorParserState();
+			rps.LinkFunction(
+				"foobar",
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number,
+				ValueType.Number
+			);
+
+			var state = new ParserState("foobar(82, 164, 329, 659) / 2", 4, rps);
+			var bindings = new RumorBindings();
+			bindings.Bind<double, double, double, double, double>(
+				"foobar", (a, b, c, d) => { return a + b + c + d; }
+			);
+
+			var n = Compiler.Math(state);
+			Assert.AreEqual(
+				new NumberValue(617),
+				n.Evaluate(new RumorScope(), bindings)
 			);
 		}