+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * Allow `jump` or `call` when specifying a default choice for `choose`.
+* Allow function calls that return a Number type in expressions.
 
 ### Changed
 * Internally, `ControlNode`s no longer store node blocks; the children nodes are